Resolve #69, New settings for high activity and hot day extra water goal factors have been added. NSNumber variables in model class have been turned to particular types.

+++ Aquaz/Controllers/IntakeViewController.swift
@@ -85,7 +85,7 @@ class IntakeViewController: UIViewController {
   }
   
   private func getInitialAmount() -> Double {
-    return intake?.amount.doubleValue ?? Double(drink.recentAmount.amount)
+    return intake?.amount ?? drink.recentAmount.amount
   }
   
   private func setupApplyButton() {
